+++ src/Admin/LicenseManagement/ApiCommunicator.php
@@ -5,7 +5,6 @@ namespace WP_SMS\Admin\LicenseManagement;
 use Exception;
 use WP_SMS\Components\RemoteRequest;
 use WP_SMS\Exceptions\LicenseException;
-use WP_SMS\Utils\AdminHelper;
 use WP_SMS\Traits\TransientCacheTrait;
 use WP_SMS\Admin\LicenseManagement\Plugin\PluginHelper;
 
@@ -147,7 +146,7 @@ class ApiCommunicator
      */
     public function validateLicense($licenseKey, $product = false)
     {
-        if (empty($licenseKey) || !AdminHelper::isStringLengthBetween($licenseKey, 32, 40) || !preg_match('/^[a-zA-Z0-9-]+$/', $licenseKey)) {
+        if (empty($licenseKey) || !preg_match('/^[a-zA-Z0-9-]+$/', $licenseKey)) {
             throw new LicenseException(
                 esc_html__('License key is not valid. Please enter a valid license and try again.', 'wp-sms'),
                 'invalid_license'